Split motion_blur into focused submodules
Convert motion_blur.rs (~2800 lines) into a motion_blur/ module with
the four concerns separated:

- schedule.rs   — SubsampleSchedule + stamp timing
- accumulator.rs — SensorAccumulator electron accumulation
- render.rs     — the render pipeline (the renamed original)
- metadata.rs   — metadata.json assembly
- mod.rs        — module docs + public re-exports

Public API of crate::sims::motion_blur is unchanged (re-exported from
mod.rs), so no downstream callers change. Each moved concern's unit
tests travel with its submodule. Pure code organization, no behavior
change.

Closes #118
